Fix error message in advanced settings overlapping 'save' button
The save button is now fully functional again when an error message
is shown.

After an invalid value is entered in the advanced settings dialog, the GUI
label for the error message that is shown was partly overlapping the
'save' button, so that the top half of the button could not be clicked
on.